+++ aspecd/processing.py
@@ -780,7 +780,8 @@ class Normalisation(SingleProcessingStep):
 
     * area
 
-      Data are divided by the sum of their *absolute* values
+      Data are divided by the sum of their *absolute* values, the number of
+      points is also taken into account.
 
     You can set these kinds using the attribute :attr:`parameters["kind"]`.
 
@@ -988,6 +989,8 @@ class Normalisation(SingleProcessingStep):
             ) - self._noise_amplitude
             self.dataset.data.axes[-1].unit = ""
         elif "area" in self.parameters["kind"].lower():
+            # might be written better
+            self.dataset.data.data /= np.sum(np.abs(data)) / data.shape[0]
             self.dataset.data.data /= np.sum(np.abs(data))
             self.dataset.data.axes[-1].unit = ""
         else:
@@ -1509,9 +1512,7 @@ class SliceExtraction(SingleProcessingStep):
         for axis in self.parameters["axis"]:
             if axis > self.dataset.data.data.ndim - 1:
                 # pylint: disable=consider-using-f-string
-                raise IndexError(
-                    "Axis %i out of bounds" % self.parameters["axis"]
-                )
+                raise IndexError("Axis %i out of bounds" % axis)
         self.parameters["unit"] = self.parameters["unit"].lower()
         if self.parameters["unit"] not in ["index", "axis"]:
             raise ValueError("Wrong unit, needs to be either index or axis.")
@@ -1826,6 +1827,9 @@ class RangeExtraction(SingleProcessingStep):
 
     .. versionadded:: 0.2
 
+    .. versionchanged:: 0.9.2
+        Range extraction with axis values sets correct upper boundary
+
 
     Examples
     --------
@@ -1871,7 +1875,7 @@ class RangeExtraction(SingleProcessingStep):
              range: [5, 10, 2]
 
     This is equivalent to ``data[5:10:2]`` or ``data[(slice(5, 10, 2))]``,
-    accordingly.
+    accordingly. Note that in Python, ranges *exclude* the upper limit.
 
     Sometimes, it is more convenient to give ranges in axis values rather
     than indices. This can be achieved by setting the parameter ``unit`` to
@@ -1888,7 +1892,10 @@ class RangeExtraction(SingleProcessingStep):
 
     Note that in this case, setting a step is meaningless and will be
     silently ignored. Furthermore, the nearest axis values will be used for
-    the range.
+    the range. Furthermore, for more intuitive use, the given range
+    *includes* the upper limit, in contrast to using indices. This is to be
+    consistent with Python's handling of ranges as weell as  with the
+    intuition of most scientists regarding the ranges for axis values.
 
     In some cases you may want to extract a range by providing percentages
     instead of indices or axis values. Even this can be done:
@@ -1959,7 +1966,7 @@ class RangeExtraction(SingleProcessingStep):
                         self.dataset.data.axes[dim].values,
                         self.parameters["range"][dim][1],
                     )
-                    slice_ = slice(start, stop)
+                    slice_ = slice(start, stop + 1)
                 else:
                     start = math.ceil(
                         self.dataset.data.axes[dim].values.size
@@ -1974,7 +1981,7 @@ class RangeExtraction(SingleProcessingStep):
                         )
                         + 1
                     )
-                    slice_ = slice(start, stop)
+                    slice_ = slice(start, stop + 1)
             # Important: Change axes first, then data
             self.dataset.data.axes[dim].values = self.dataset.data.axes[
                 dim